role base working

+++ app/Http/Controllers/API/UserController.php
@@ -2,13 +2,25 @@
 
 namespace App\Http\Controllers\API;
 
-use App\Http\Controllers\Controller;
-use App\Models\User;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use App\User;
 use Illuminate\Support\Facades\Hash;
+
 
 class UserController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,7 +28,11 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        // $this->authorize('isAdmin');
+        if (\Gate::allows('isAdmin') || \Gate::allows('isAuthor')) {
+            return User::latest()->paginate(5);
+        }
+
     }
 
     /**
@@ -27,10 +43,11 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
+
+        $this->validate($request,[
             'name' => 'required|string|max:191',
-            'email' => 'required|string|max:191|unique:users,email',
-            'password' => 'required|min:6'
+            'email' => 'required|string|email|max:191|unique:users',
+            'password' => 'required|string|min:6'
         ]);
 
         return User::create([
@@ -41,9 +58,13 @@ class UserController extends Controller
             'photo' => $request['photo'],
             'password' => Hash::make($request['password']),
         ]);
+
+
     }
 
-    public function updateProfile(Request $request){
+
+    public function updateProfile(Request $request)
+    {
         $user = auth('api')->user();
 
 
@@ -81,7 +102,8 @@ class UserController extends Controller
     }
 
 
-    public function profile(){
+    public function profile()
+    {
         return auth('api')->user();
     }
 
@@ -93,7 +115,7 @@ class UserController extends Controller
      */
     public function show($id)
     {
-
+        //
     }
 
     /**
@@ -105,17 +127,17 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+
         $user = User::findOrFail($id);
 
-        $this->validate($request, [
+        $this->validate($request,[
             'name' => 'required|string|max:191',
-            'email' => 'required|string|max:191|unique:users,email,'.$user->id,
+            'email' => 'required|string|email|max:191|unique:users,email,'.$user->id,
             'password' => 'sometimes|min:6'
         ]);
 
         $user->update($request->all());
-
-        return ['message' => 'Updated user info'];
+        return ['message' => 'Updated the user info'];
     }
 
     /**
@@ -126,10 +148,29 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+
+        $this->authorize('isAdmin');
+
         $user = User::findOrFail($id);
+        // delete the user
 
         $user->delete();
 
         return ['message' => 'User Deleted'];
+    }
+
+    public function search(){
+
+        if ($search = \Request::get('q')) {
+            $users = User::where(function($query) use ($search){
+                $query->where('name','LIKE',"%$search%")
+                        ->orWhere('email','LIKE',"%$search%");
+            })->paginate(20);
+        }else{
+            $users = User::latest()->paginate(5);
+        }
+
+        return $users;
+
     }
 }
